feat: centralize sandbox bootstrap
